+++ client/components/product/product-card2.js
@@ -1,6 +1,6 @@
 import { Fragment, useState } from "react";
 import { CiHeart } from "react-icons/Ci";
-import { BiSolidHeart,BiSolidCart} from "react-icons/bi";
+import { BiSolidHeart, BiSolidCart } from "react-icons/bi";
 import axios from 'axios';
 import { useEffect } from "react";
 
@@ -13,56 +13,57 @@ export default function ProductCard2() {
 
 
   //讀取資料庫資料
-  const [productData, setProductData] = useState({}); // 用來存放從後端獲取的產品資料
+  const [productData, setProductData] = useState({ result: [] }); // 初始化為一個帶有 result 屬性的物件
 
   useEffect(() => {
-    // 在組件加載後，發送 GET 請求到後端 API 以獲取產品資料
     axios.get("http://localhost:3005/api/product").then((response) => {
-      // response.data 包含了後端返回的 JSON 資料
-      setProductData(response.data.result);
+      setProductData({ result: response.data.result }); // 將伺服器端的 result 放入物件中
     });
-  }, []); // 空數組表示這個效果只會在組件加載時執行一次
-
-
+  }, []);
 
   return (
     <>
-      <div className="product-card">
-        {/* <div className="card" style={{ width: "275px", height: "425px" }}> */}
-          <div className="card" >
-          <img
-            src="https://cdn-front.mao-select.com.tw//upload_files/fonlego-rwd/prodpic/D_M3PD150101-e-0.jpg"
-            className="card-img-top"
-            alt="..."
-          />
-          <div className="card-body">
-            <div className="d-flex justify-content-between align-items-center">
-            {/* 類別按鈕顏色已建好 btn-color-1 一直到btn-color-7 再依需求調整className即可 */}
-            {/* 顏色設定如果需要再調整，可以到以下檔案調整 \final-project-pet\client\styles\components-style\_product-card.scss */}
-              <div className="btn btn-color-1 d-flex align-items-center">貓草</div>
-              <div className="" onClick={toggleFavorite}>
-                {isFavorite ? (
-                  <div className="position-relative inline-block">
-                    <BiSolidHeart className="heart-icon-red" />
-                    <CiHeart className="heart-icon-red-line" />
-                  </div>
-                ) : (
-                  <CiHeart />
-                )}
-              </div>
-            </div>
+      {productData.result.map((v, i) => {
+        return (
+          <div className="col-6 col-md-4 col-lg-6 col-xl-4">
+            <div className="product-card2" key={v.product_id}>
+              <div className="card" >
+                
 
-            <div className="card-text size-6 m-size-7">{productData.product_name}</div>
-            <div className="product-sale-price d-flex align-items-center" href="#">
-              <del>$220</del>
-              <div className="price fs-4 mx-2  size-7 m-size-7">NT$140</div>
-              <div className="rounded-circle cart p-1 d-flex justify-content-end align-items-center">
-              <BiSolidCart className="cart"/>
+                <img
+                  src={v.images_one}
+                  className="card-img-top"
+                  alt="..."
+                />
+
+               
+
+                <div className="card-body" >
+                  <div className="d-flex justify-content-between align-items-center">
+                    {/* 類別按鈕顏色已建好 btn-color-1 一直到btn-color-7 再依需求調整className即可 */}
+                    {/* 顏色設定如果需要再調整，可以到以下檔案調整 \final-project-pet\client\styles\components-style\_product-card.scss */}
+                    <div className="btn btn-color-1 d-flex align-items-center py-3">{v.subcategory_name}</div>
+                  </div>
+
+                  <div className="card-text size-6 m-size-7">{v.product_name}</div>
+                  <div className="product-sale-price d-flex align-items-center" href="#">
+                    <div className="price fs-4 mx-2 size-7 m-size-7">NT${v.specialoffer}</div>
+                    <del>NT${v.price}</del>
+                  </div>
+                  <button type="submit" className=" btn-confirm mx-1">
+                    <BiSolidCart />
+                  </button>
+                </div>
               </div>
             </div>
           </div>
-        </div>
-      </div>
+
+        )
+      })}
+
     </>
   );
 }
+
+
+
