+++ client/pages/product/index.js
@@ -4,6 +4,8 @@ import Search from '@/components/job/search';
 import ProductCard2 from '@/components/product/product-card2';
 import { FaCaretUp, FaCaretDown } from 'react-icons/fa';
 import ProductListOffcanvas from '@/components/product/product-list-offcanvas';
+import { useEffect } from 'react';
+import axios from 'axios';
 
 
 export default function ProductList() {
@@ -29,7 +31,14 @@ export default function ProductList() {
         }
     };
 
-    
+    const [subcategoryData, setSubcategoryData] = useState({ result: [] });
+    useEffect(() => {
+        axios.get("http://localhost:3005/api/product/category/subcategory").then((response) => {
+            setSubcategoryData({ result: response.data.result });
+        });
+    }, [])
+
+
 
     return (
         <>
@@ -58,120 +67,108 @@ export default function ProductList() {
                     <div className="product-offcanvas-m d-block d-lg-none ">
                         <ProductListOffcanvas />
                     </div>
-                    <section className='sidebar-product d-flex  flex-column flex-lg-row mt-3 '>
-                        <div className='sidebar col-md-3 ms-3 me-1 d-none d-lg-block pe-4'>
-                            <div className="accordion" id="accordionPanelsStayOpenExample">
-                                <div className="accordion-item">
-                                    <h2 className="accordion-header" id="panelsStayOpen-headingCategory">
-                                        <button
-                                            className="accordion-button "
-                                            type="button"
-                                            data-bs-toggle="collapse"
-                                            data-bs-target="#panelsStayOpen-collapseCategory"
-                                            aria-expanded="true"
-                                            aria-controls="panelsStayOpen-collapseCategory"
-                                        >
-                                            大類1
-                                        </button>
-                                    </h2>
-                                    <div id="panelsStayOpen-collapseCategory" class="accordion-collapse collapse show">
-                                        <div className="accordion-body row">
-                                            <button className="button-subcategory" type="button" >
-                                                小類
-                                            </button>
-                                            <button className="button-subcategory" type="button row">
-                                                小類
-                                            </button>
-                                        </div>
-                                    </div>
-                                    <div className="accordion-item">
-                                        {/* 要記得改id/data-bs-target/aria-controls/aria-labelledby數字 */}
-                                        <h2 className="accordion-header" id="panelsStayOpen-headingTwo">
-                                            <button className="accordion-button collapsed " type="button" data-bs-toggle="collapse" data-bs-target="#panelsStayOpen-collapseTwo" aria-expanded="false" aria-controls="panelsStayOpen-collapseTwo">
-                                                大類2
-                                            </button>
-                                        </h2>
-                                        <div id="panelsStayOpen-collapseTwo" class="accordion-collapse collapse" aria-labelledby="panelsStayOpen-headingTwo">
-                                            <div className="accordion-body row">
-                                                <button className="button-subcategory no-border-btn" type="button" >
-                                                    小類
+
+                    { subcategoryData.result.map((v,i)=>{
+                        return (
+                            <section className='sidebar-product d-flex  flex-column flex-lg-row mt-3' key={v.category_id}>
+                                <div className='sidebar col-md-3 ms-3 me-1 d-none d-lg-block pe-4'>
+                                    <div className="accordion" id={`accordionPanelsStayOpenExample-${v.category_id}`}>
+                                        <div className="accordion-item">
+                                            <h2 className="accordion-header" id={`panelsStayOpen-headingCategory-${v.category_id}`}>
+                                                <button
+                                                    className="accordion-button"
+                                                    type="button"
+                                                    data-bs-toggle="collapse"
+                                                    data-bs-target={`#panelsStayOpen-collapseCategory-${v.category_id}`}
+                                                    aria-expanded="true"
+                                                    aria-controls={`panelsStayOpen-collapseCategory-${v.category_id}`}
+                                                >
+                                                    {v.category_name}
                                                 </button>
-                                                <button className="button-subcategory" type="button">
-                                                    小類
-                                                </button>
-                                            </div>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className='filter mt-3 '>
-                                <div className="card filter-card">
-                                    <div className="card-header">
-                                        其他選項
-                                    </div>
-                                    <div className="card-body">
-                                        <form>
-                                            <div className="col-12">
-                                                <label for="inputprice" className="form-label">價格區間</label>
-                                                <div className="row col-md">
-                                                    <div className="col-md-5">
-                                                        <input type="number" className="form-control" id="price" placeholder="$最低價">
-                                                        </input>
-                                                    </div>
-                                                    <div class="col-md dash">
-                                                        ~
-                                                    </div>
-                                                    <div className="col-md-5">
-                                                        <input type="number" className="form-control" id="price" placeholder="$最高價">
-                                                        </input>
-                                                    </div>
+                                            </h2>
+                                            <div id={`panelsStayOpen-collapseCategory-${v.category_id}`} className="accordion-collapse collapse show" aria-labelledby={`panelsStayOpen-headingCategory-${v.category_id}`}>
+                                                <div className="accordion-body row">
+                                                    {v.subcategories && v.subcategories.map((subcategory, j) => (
+                                                        <button className="button-subcategory" type="button" key={subcategory.subcategory_id}>
+                                                            {subcategory.subcategory_name}
+                                                        </button>
+                                                    ))}
                                                 </div>
                                             </div>
-                                            <div className="col-12 mt-2">
-                                                <label for="brand" className="form-label">品牌</label>
-                                                <input type="text" className="form-control" id="brand" placeholder="請輸入品牌關鍵字">
-                                                </input>
+                                        </div>
+                                        <div className="accordion-item">
+                                            <h2 className="accordion-header" id={`panelsStayOpen-headingTwo-${v.category_id}`}>
+                                                <button className="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target={`#panelsStayOpen-collapseTwo-${v.category_id}`} aria-expanded="false" aria-controls={`panelsStayOpen-collapseTwo-${v.category_id}`}>
+                                                    {v.category_name}
+                                                </button>
+                                            </h2>
+                                            <div id={`panelsStayOpen-collapseTwo-${v.category_id}`} className="accordion-collapse collapse" aria-labelledby={`panelsStayOpen-headingTwo-${v.category_id}`}>
+                                                <div className="accordion-body row">
+                                                    {v.subcategories && v.subcategories.map((subcategory, j) => (
+                                                        <button className="button-subcategory no-border-btn" type="button" key={subcategory.subcategory_id}>
+                                                            {subcategory.subcategory_name}
+                                                        </button>
+                                                    ))}
+                                                </div>
                                             </div>
-                                            <button type="submit" className="btn btn-brown col-12 mt-3">
-                                                確定
-                                            </button>
-                                        </form>
+                                        </div>
+                                    </div>
+
+
+
+                                    <div className='filter mt-3 '>
+                                        <div className="card filter-card">
+                                            <div className="card-header">
+                                                其他選項
+                                            </div>
+                                            <div className="card-body">
+                                                <form>
+                                                    <div className="col-12">
+                                                        <label for="inputprice" className="form-label">價格區間</label>
+                                                        <div className="row col-md">
+                                                            <div className="col-md-5">
+                                                                <input type="number" className="form-control" id="price" placeholder="$最低價">
+                                                                </input>
+                                                            </div>
+                                                            <div class="col-md dash">
+                                                                ~
+                                                            </div>
+                                                            <div className="col-md-5">
+                                                                <input type="number" className="form-control" id="price" placeholder="$最高價">
+                                                                </input>
+                                                            </div>
+                                                        </div>
+                                                    </div>
+                                                    <div className="col-12 mt-2">
+                                                        <label for="brand" className="form-label">品牌</label>
+                                                        <input type="text" className="form-control" id="brand" placeholder="請輸入品牌關鍵字">
+                                                        </input>
+                                                    </div>
+                                                    <button type="submit" className="btn btn-brown col-12 mt-3">
+                                                        確定
+                                                    </button>
+                                                </form>
+                                            </div>
+                                        </div>
                                     </div>
                                 </div>
-                            </div>
-                        </div>
-                        <div className='product d-lg-flex flex-column justify-content-center '>
-                            <div className=" row d-flex mb-3 g-3 g-md-4 ">
-                                <div className="col-6 col-md-4 col-lg-6 col-xl-4">
-                                    <ProductCard2 />
-                                </div>
-                                <div className="col-6 col-md-4 col-lg-6 col-xl-4">
-                                    <ProductCard2 />
-                                </div>
-                                <div className="col-6 col-md-4 col-lg-6 col-xl-4">
-                                    <ProductCard2 />
-                                </div>
-                                <div className="col-6 col-md-4 col-lg-6 col-xl-4">
-                                    <ProductCard2 />
-                                </div>
-                                <div className="col-6 col-md-4 col-lg-6 col-xl-4">
-                                    <ProductCard2 />
-                                </div>
-                                <div className="col-6 col-md-4 col-lg-6 col-xl-4">
-                                    <ProductCard2 />
-                                </div>
-                                <div className="col-6 col-md-4 col-lg-6 col-xl-4">
-                                    <ProductCard2 />
-                                </div>
-                                
-                                {/* 原本錯的 */}
-                                {/* <div className="col-lg-4 col-md-4 col-sm-6">
+
+
+
+                                <div className='product d-lg-flex flex-column justify-content-center '>
+                                    <div className=" row d-flex mb-3 g-3 g-md-4 ">
+                                        <ProductCard2 />
+                                        {/* 原本錯的 */}
+                                        {/* <div className="col-lg-4 col-md-4 col-sm-6">
                                     <ProductCard />
                                 </div> */}
-                            </div>
-                        </div>
+                                    </div>
+                                </div>
 
-                    </section>
+                            </section>
+
+                        )
+                    })}
 
                 </div>
             </div>
